index fillna tests
