+++ pandas/tests/indexes/datetimes/test_fillna.py
